feat(core): remove plugin install / disable / enable logic
moves initialize logic into plugin_manager
closes #872
